mdb: implement commit API.
Idea is to be able to commit a latest batch of mutations into an
existing index snapshot, there by creating a new snapshot. It is
a pre-requisite that the new batch of mutation and its seqno must
all be newer than the index snapshot into which it is being
committed.