mdb::OMap refactor and breaking change.
Now is this more inline with other variants of OMap with just one
feature, which is, it is partially persistent. That is all writes are
serialized and all the threads tend to access the same root using
a spin-lock and holds a read reference.

TODO: figure out a proper test case for this variant.